+++ frontend/react-frontend/src/components/Nav.jsx
@@ -9,13 +9,13 @@ const Nav = () => {
     <nav>
       <h2>DevRoast</h2>
       <ul>
-        <Link to="/">
-          <li>Home</li>
-        </Link>
-        <Link to="/login">
-          <li>Login</li>
-        </Link>
-          <li>{user['key']}</li>
+        <Link to="/"> <li>Home</li> </Link>
+        {user['key'] ? 
+        <>
+          <li>{user['key']} is Logged In </li> 
+          <Link to="/"> <li onClick={() => setUser({...user, 'key':null})}> Log Out </li> </Link> 
+        </>
+        : <Link to="/login"> <li>Log In</li> </Link> }
       </ul>
     </nav>
   );
